+++ tests/image_methods_test.py
@@ -1197,6 +1197,19 @@ def test_kuwahara():
         assert was != img.signature
 
 
+def test_label():
+    with Image(filename='rose:') as img:
+        was = img.signature
+        img.label('a', left=0, top=0, font=Font('monospace', 12))
+        now = img.signature
+        assert now != was
+        img.label('b', font=Font('monospace', 12), gravity='south')
+        assert img.signature != now
+    with raises(TypeError):
+        with Image(filename='rose:') as img:
+            img.label('x')
+
+
 def test_level(fx_asset):
     with Image(filename=str(fx_asset.join('gray_range.jpg'))) as img:
         # Adjust the levels to make this image entirely black